[Rails 3.1] Improved search API and corrected a few bugs.
- Added a logger. It defaults to Rails.logger or STDOUT when using the Rails
console, but can be overriden in initializers.
- Added a configuration class to hold not only the CloudSearch domain, but
also information about index fields, in order to be able to properly translate
search parameters.
- Changed the model callback methods to allow for application models to
override the callbacks, if needed.
- Improved search API: the API now looks a little more like Model.where(....),
in that it's possible to pass a hash of model attributes and values that will
be used as filters in the search.
- Improved search API: the cloud_search method now supports pagination and
sorting done directly in CloudSearch
- Bug fix: added require 'tinia/railtie' so that it's not necessary to call
Tinia.activate_active_record! in Rails 3.
- Bug fix: using Time.now.utc instead of updated_at as the CloudSearch
document version. This was causing documents not to be deleted and inserted
correctly.